pyproject.toml
[project]
name = "react-fastapi-admin"
version = "0.1.0"
description = "React Fastapi admin"
requires-python = ">=3.11"
dependencies = [
    "aerich==0.9.2",
    "aiosqlite==0.22.1",
    "bcrypt==5.0.0",
    "email-validator==2.3.0",
    "fastapi==0.135.1",
    "granian[reload]==2.7.2",
    "loguru==0.7.3",
    "pydantic==2.12.5",
    "pydantic-settings==2.13.1",
    "pytest==8.3.5",
    "pyjwt==2.12.1",
    "python-dotenv==1.2.2",
    "python-multipart==0.0.22",
    "starlette==0.52.1",
    "tortoise-orm==1.1.6",
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv]
override-dependencies = ["tortoise-orm==1.1.6"]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.black]
line-length = 120
target-version = ["py311"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
extend-select = []
ignore = ["F403", "F405"]

[tool.isort]
profile = "black"
line_length = 120

[tool.aerich]
tortoise_orm = "app.settings.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
